ok1

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: fpolini <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/27 18:15:33 by fpolini           #+#    #+#             */
-/*   Updated: 2015/11/29 13:45:33 by fpolini          ###   ########.fr       */
+/*   Updated: 2015/12/02 13:31:36 by fpolini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int		ft_strlcat(char *dst, char *src, size_t n)
 {
-	int i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
