+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: fpolini <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/30 12:32:39 by fpolini           #+#    #+#             */
-/*   Updated: 2015/12/01 17:33:39 by fpolini          ###   ########.fr       */
+/*   Updated: 2015/12/02 13:27:48 by fpolini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char		*ft_strnstr(const char *s1, const char *s2, size_t n)
 {
-	const char	*pointeur;
-	int			i;
+	const char		*pointeur;
+	size_t			i;
 
 	pointeur = &s1[0];
 	i = 0;
